Programm/requirements.txt
numpy
matplotlib
pyQt5
opencv-python
scikit-image
pywin32
mysql-connector-python